TrabajoFinal2023-2/requirements.txt
tqdm==4.66.1
loguru==0.7.2
reportlab==4.0.5
python-docx==0.8.11
python-pptx==0.6.22
XlsxWriter==3.1.2
beautifulsoup4==4.11.2
bokeh==3.2.2
datascience==0.17.6
db-dtypes==1.1.1
Flask==2.2.5
folium==0.14.0
geopandas==0.13.2
geopy==2.3.0
kaggle==1.5.16
keras==2.12.0
lxml==4.9.3
Markdown==3.4.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.1
matplotlib-inline==0.1.6
matplotlib-venn==0.11.9
numpy==1.23.5
opencv-contrib-python==4.8.0.76
opencv-python==4.8.0.76
opencv-python-headless==4.8.0.76
openpyxl==3.1.2
pandas==1.5.3
pandas-datareader==0.10.0
pandas-gbq==0.17.9
Pillow==9.4.0
pip-tools==6.13.0
plotly==5.15.0
plotnine==0.12.2
polars==0.17.3
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
requirements-parser==0.5.0
scikit-image==0.19.3
scikit-learn==1.2.2
scipy==1.10.1
scs==3.2.3
seaborn==0.12.2
shapely==2.0.1
sklearn-pandas==2.2.0
SQLAlchemy==2.0.20
sqlparse==0.4.4
statsmodels==0.14.0
sympy==1.12
tensorboard==2.12.3
tensorboard-data-server==0.7.1
tensorflow==2.12.0
tensorflow-datasets==4.9.2
tensorstore==0.1.41
urllib3==2.0.4